Попровки после первого ревью 1. Типизировал хук useForm 2. Убрал any в forgot-password-form.tsx
